requirements.txt
streamlit>=1.29.0
torch==1.13.1
torchvision==0.14.1
timm==0.6.12
matplotlib==3.7.1
numpy==1.24.3
pandas==1.5.3
Pillow==9.5.0
scikit-learn==1.2.2
tqdm==4.65.0
python-dotenv==1.0.0
plotly==5.14.1 